Revert "Revert "Make `parseEnv` a helper function""
This reverts commit b87ba2f046165f2f25503978ba1e0292d02c0d45.
